Create README.md
PythonWeb is a Python-based web development tool.
To install it,
install Python.
Next, enter "pip install PyQt5 requests" at the command prompt.
Once the installation is complete, enter "Python (where the path to PythonWeb)" at the command prompt. If it runs, the installation was successful.
Thank you for reading to the end.